Standarize wasmer_wasm.h and wasmer_deprecated.h into wasmer.h
