+++ lib/c-api/src/wasm_c_api/module.rs
@@ -30,7 +30,7 @@ pub struct wasm_module_t {
 /// # use inline_c::assert_c;
 /// # fn main() {
 /// #    (assert_c! {
-/// # #include "tests/wasmer_wasm.h"
+/// # #include "tests/wasmer.h"
 /// #
 /// int main() {
 ///     wasm_engine_t* engine = wasm_engine_new();
@@ -183,7 +183,7 @@ mod tests {
     #[test]
     fn test_module_validate() {
         (assert_c! {
-            #include "tests/wasmer_wasm.h"
+            #include "tests/wasmer.h"
 
             int main() {
                 wasm_engine_t* engine = wasm_engine_new();
@@ -210,7 +210,7 @@ mod tests {
     #[test]
     fn test_module_new() {
         (assert_c! {
-            #include "tests/wasmer_wasm.h"
+            #include "tests/wasmer.h"
 
             int main() {
                 wasm_engine_t* engine = wasm_engine_new();
@@ -239,7 +239,7 @@ mod tests {
     #[test]
     fn test_module_exports() {
         (assert_c! {
-            #include "tests/wasmer_wasm.h"
+            #include "tests/wasmer.h"
 
             int main() {
                 wasm_engine_t* engine = wasm_engine_new();
@@ -345,7 +345,7 @@ mod tests {
     #[test]
     fn test_module_imports() {
         (assert_c! {
-            #include "tests/wasmer_wasm.h"
+            #include "tests/wasmer.h"
 
             int main() {
                 wasm_engine_t* engine = wasm_engine_new();
@@ -461,7 +461,7 @@ mod tests {
     #[test]
     fn test_module_serialize() {
         (assert_c! {
-            #include "tests/wasmer_wasm.h"
+            #include "tests/wasmer.h"
 
             int main() {
                 wasm_engine_t* engine = wasm_engine_new();
@@ -493,7 +493,7 @@ mod tests {
     #[test]
     fn test_module_serialize_and_deserialize() {
         (assert_c! {
-            #include "tests/wasmer_wasm.h"
+            #include "tests/wasmer.h"
 
             int main() {
                 wasm_engine_t* engine = wasm_engine_new();
