+++ lib/c-api/tests/test-wat2wasm.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <inttypes.h>
 
-#include "wasmer_wasm.h"
+#include "wasmer.h"
 
 #define own
