+++ lib/cli/src/commands/wasmer_create_exe_main.c
@@ -2,8 +2,7 @@
 extern "C" {
 #endif
 
-#include "wasmer_wasm.h"
-#include "wasm.h"
+#include "wasmer.h"
 #include "my_wasm.h"
 
 #include <stdio.h>
