+++ tests/integration/cli/tests/object_file_engine_test_c_source.c
@@ -2,8 +2,7 @@
 extern "C" {
 #endif
 
-#include "wasmer_wasm.h"
-#include "wasm.h"
+#include "wasmer.h"
 #include "my_wasm.h"
 
 #include <stdio.h>
